+++ spec/is_it_done_yet_spec.rb
@@ -1,7 +1,7 @@
-require "spec_helper"
+require 'spec_helper'
 
 RSpec.describe IsItDoneYet do
-  it "has a version number" do
+  it 'has a version number' do
     expect(IsItDoneYet::VERSION).not_to be nil
   end
 end